requirements.txt
fastapi
gunicorn
sqlalchemy
pyodbc
python-jose[cryptography]
passlib[bcrypt]
requests
PyPDF2
openai
pydantic
python-dotenv
